Moved main.py into answer_bear package and exposed app object.
